Anpassung: int und float wurden durch np.integer bzw. np.float ersetzt

+++ tensor/htucker/_add_and_truncate.py
@@ -28,17 +28,17 @@ def add_and_truncate(cls, summanden, max_rank, abs_err=None, rel_err=None, copy=
         raise ValueError("Nicht alle hierarchische Tuckertensoren aus 'summanden' haben dieselben Modi.")
     if not all(equal(element.dtree, summanden[0].dtree) for element in summanden):
         raise ValueError("Nicht alle hierarchische Tuckertensoren aus 'summanden' haben identische Dimensionsbäume.")
-    if not np.issubdtype(type(max_rank), int):
+    if not np.issubdtype(type(max_rank), np.integer):
         raise TypeError("'max_rank' muss ein int>= 1 sein.")
     if not max_rank >= 1:
         raise ValueError("'max_rank' muss ein int>= 1 sein.")
     if abs_err is not None:
-        if not np.issubdtype(type(abs_err), float):
+        if not np.issubdtype(type(abs_err),np.float):
             raise TypeError("'abs_err' muss ein float > 0 sein.")
         if not abs_err > 0:
             raise ValueError("'abs_err' muss ein float > 0 sein.")
     if rel_err is not None:
-        if not np.issubdtype(type(rel_err), float):
+        if not np.issubdtype(type(rel_err),np.float):
             raise TypeError("'rel_err' muss ein float > 0 sein.")
         if not rel_err > 0:
             raise ValueError("'rel_err' muss ein float > 0 sein.")
