+++ tensor/htucker/_trunc_rank.py
@@ -15,17 +15,17 @@ def trunc_rank(s, max_rank, abs_err=None, rel_err=None):
         raise ValueError("'s' muss ein positiver 1D-np.ndarray mit mindestens einem Eintrag sein.")
     if not all(s.shape[i] > 0 for i in range(len(s.shape))):
         raise ValueError("'s' muss ein positiver 1D-np.ndarray mit mindestens einem Eintrag sein.")
-    if not np.issubdtype(type(max_rank), int):
+    if not np.issubdtype(type(max_rank), np.integer):
         raise TypeError("'max_rank' muss ein int mit 'max_rank'>= 1 sein.")
     if not max_rank >= 1:
         raise ValueError("'max_rank' muss ein int mit 'max_rank'>= 1 sein.")
     if abs_err is not None:
-        if not np.issubdtype(type(abs_err), float):
+        if not np.issubdtype(type(abs_err),np.float):
             raise TypeError("'abs_err' muss ein positiver float sein.")
         if not abs_err > 0:
             raise ValueError("'abs_err' muss ein positiver float sein.")
     if rel_err is not None:
-        if not np.issubdtype(type(rel_err), float):
+        if not np.issubdtype(type(rel_err),np.float):
             raise TypeError("'rel_err' muss ein positiver float sein.")
         if not rel_err > 0:
             raise ValueError("'rel_err' muss ein positiver float sein.")
